'comment'

+++ database/migrations/2020_09_06_100338_create_clients_table.php
@@ -17,7 +17,7 @@ class CreateClientsTable extends Migration
             $table->increments('id');
             $table->text('facebook');
             $table->integer('points');
-            $table->integer('fb_id');
+            $table->bigInteger('fb_id');
             $table->timestamps();
             $table->text('slug');
